Builtin history (#120)
[Builtin][History]: add builtin history

This builtin lists previous shell commands with their indexes.

Usage : history [-nr] [first [last]]

Use -n to hide indexes in the display.
Use -r to reverse list order.
`first` and `last` have the following form:
[+]<index> is an absolute index.
-<index> references the last nth command (starts from -1).
<string> references the first command that starts with <string>
If `last` is not given, lists to the last command. If `first` neither, lists all history.

Notes:
Indices starts from 1 (the first command in the history) and increments monotonically for each command. Eventually wraps when exceeds the maximum value of a `size_t`. In this case this is an undefined behavior.
Since the history is truncated to HISTSIZE, commands in the history may not start from index 1.