[Exec]: fix subshell error with pipe
Getting back the controlling terminal was done on STDIN_FILENO.
But in a pipe, stdin is not the terminal, but the read end of the pipe.

Apart from that bug, the signal was not set correctly, SIGTTOU, in case
of a subshell.
And finally, restoring the controlling terminal after a wait was done
with the pid of the process, instead of his pgid